+++ Arbiter.App/Views/MainWindow.axaml.cs
@@ -1,6 +1,7 @@
 using System;
 using Arbiter.App.Controls;
 using Arbiter.App.ViewModels;
+using Avalonia;
 using Avalonia.Controls;
 using Avalonia.Interactivity;
 
@@ -8,6 +9,8 @@ namespace Arbiter.App.Views;
 
 public partial class MainWindow : TalgoniteWindow
 {
+    private bool _canClose;
+    
     private MainWindowViewModel ViewModel => (DataContext as MainWindowViewModel)!;
         
     public MainWindow()
@@ -41,14 +44,29 @@ public partial class MainWindow : TalgoniteWindow
 
     private async void Window_OnClosing(object? sender, WindowClosingEventArgs e)
     {
+        if (_canClose)
+        {
+            return;
+        }
+
         try
         {
-            var canClose = await ViewModel.OnClosing(e.CloseReason);
-            e.Cancel = !canClose;
+            // Do not close the window immediately, wait for async cleanup tasks
+            // If you do not do this, the application will exit prematurely
+            e.Cancel = true;
+            _canClose = await ViewModel.OnClosing(e.CloseReason);
+
+            // Re-trigger closing the window which will now actually close
+            if (_canClose)
+            {
+                Close();
+            }
         }
         catch
         {
-            e.Cancel = false;
+            // There was an issue, the window still needs to close
+            _canClose = true;
+            Close();
         }
     }
 }